+++ src/core/component/state/index.ts
@@ -12,9 +12,13 @@
  */
 
 import watch from 'core/object/watch';
+import { State } from 'core/component/state/interface';
 
-export default watch({
+export * from 'core/component/state/interface';
+
+export default watch(<State>{
 	isAuth: undefined,
 	isOnline: undefined,
-	lastOnlineDate: undefined
+	lastOnlineDate: undefined,
+	experiments: undefined
 }).proxy;