+++ src/super/i-block/modules/block/const.ts
@@ -1,0 +1,14 @@
+/*!
+ * V4Fire Client Core
+ * https://github.com/V4Fire/Client
+ *
+ * Released under the MIT license
+ * https://github.com/V4Fire/Client/blob/master/LICENSE
+ */
+
+export const
+	fakeCtx = document.createElement('div'),
+	modRgxpCache = Object.createDict<RegExp>();
+
+export const
+	elRxp = /_+/;