+++ src/super/i-block/modules/opt/index.ts
@@ -24,18 +24,18 @@ export * from 'super/i-block/modules/opt/interface';
  */
 export default class Opt extends Friend {
 	/** @see [[iBlock.ifOnceStore]] */
-	protected get ifOnceStore(): this['CTX']['ifOnceStore'] {
+	protected get ifOnceStore(): Dictionary<number> {
 		return this.ctx.ifOnceStore;
 	}
 
 	/**
 	 * Returns a number if the specified label:
-	 *   2 -> already exists in a cache;
-	 *   1 -> just written in a cache;
-	 *   0 -> doesn't exist in a cache.
+	 *   2 -> already exists in the cache;
+	 *   1 -> just written in the cache;
+	 *   0 -> doesn't exist in the cache.
 	 *
 	 * This method is used with conditions to provide a logic: if the condition was switched to true,
-	 * then further it always return true.
+	 * then further it always returns true.
 	 *
 	 * @param label
 	 * @param [value] - label value (will be saved in the cache only if true)
@@ -50,12 +50,11 @@ export default class Opt extends Friend {
 		const
 			strLabel = String(label);
 
-		if (this.ifOnceStore[strLabel]) {
+		if (this.ifOnceStore[strLabel] != null) {
 			return 2;
 		}
 
 		if (value) {
-			// @ts-ignore (access)
 			return this.ifOnceStore[strLabel] = 1;
 		}
 
@@ -63,8 +62,8 @@ export default class Opt extends Friend {
 	}
 
 	/**
-	 * Tries to find a cached alternative to the specified value and returns it,
-	 * or if the value wasn't find in the cache, it will be frozen, cached and returned.
+	 * Tries to find a blueprint in the cache to the specified value and returns it,
+	 * or if the value wasn't found in the cache, it would be frozen, cached, and returned.
 	 *
 	 * This method is used to cache raw literals within component templates to avoid redundant re-renders that occurs
 	 * because a link to an object was changed.
@@ -76,17 +75,17 @@ export default class Opt extends Friend {
 	 * < b-button :mods = opt.memoizeLiteral({foo: 'bla'})
 	 * ```
 	 */
-	memoizeLiteral<T>(literal: T): T extends (infer V)[] ? ReadonlyArray<V> : T extends object ? Readonly<T> : T {
+	memoizeLiteral<T>(literal: T): T extends Array<infer V> ? readonly V[] : T extends object ? Readonly<T> : T {
 		if (Object.isFrozen(literal)) {
 			return <any>literal;
 		}
 
 		const key = Object.fastHash(literal);
-		return literalCache[key] = literalCache[key] || Object.freeze(<any>literal);
+		return literalCache[key] = literalCache[key] ?? Object.freeze(<any>literal);
 	}
 
 	/**
-	 * Shows any changes of the component properties.
+	 * Shows any changes of component properties.
 	 * This method is useful to debug.
 	 */
 	showAnyChanges(): void {