+++ config/CHANGELOG.md
@@ -9,6 +9,16 @@ Changelog
 > - :house:      [Internal]
 > - :nail_care:  [Polish]
 
+## v3.0.0-rc.41 (2020-07-29)
+
+#### :boom: Breaking Change
+
+* Removed redundant parameters from `snakeskin`
+
+#### :rocket: New Feature
+
+* Added `csp`
+
 ## v3.0.0-rc.36 (2020-07-13)
 
 #### :rocket: New Feature