+++ src/core/component/engines/vue/CHANGELOG.md
@@ -9,6 +9,18 @@ Changelog
 > - :house:      [Internal]
 > - :nail_care:  [Polish]
 
+## v3.0.0-rc.44 (2020-07-30)
+
+#### :bug: Bug Fix
+
+* Fixed setting of `staticClass`
+
+## v3.0.0-rc.40 (2020-07-27)
+
+#### :house: Internal
+
+* Logging Vue errors and warnings via the `core/log` module
+
 ## v3.0.0-rc.37 (2020-07-20)
 
 #### :house: Internal