+++ src/core/component/state/CHANGELOG.md
@@ -9,6 +9,13 @@ Changelog
 > - :house:      [Internal]
 > - :nail_care:  [Polish]
 
+## v3.0.0-rc.46 (2020-07-31)
+
+#### :house: Internal
+
+* Added `experiments`
+* Added `interface/State`
+
 ## v3.0.0-rc.37 (2020-07-20)
 
 #### :bug: Bug Fix