+++ src/super/i-static-page/CHANGELOG.md
@@ -9,6 +9,28 @@ Changelog
 > - :house:      [Internal]
 > - :nail_care:  [Polish]
 
+## v3.0.0-rc.43 (2020-07-30)
+
+#### :bug: Bug Fix
+
+* Fixed generation of `init.js`
+
+## v3.0.0-rc.41 (2020-07-29)
+
+#### :boom: Breaking Change
+
+* Removed `async`, `module`, `nomodule` from `modules/interface.js/Lib`
+
+#### :rocket: New Feature
+
+* Added `links` to `deps.js`
+* Added `attrs` to `modules/interface.js`
+* Added generation of `init.js`
+
+#### :house: Internal
+
+* Moved logic from SS to JS
+
 ## v3.0.0-rc.37 (2020-07-20)
 
 #### :house: Internal